requirements.txt
fastapi==0.121.2
uvicorn==0.21.1
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic[email]==2.12.4
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.3.0
celery==5.2.7
redis==4.5.4
python-dotenv==1.0.0
python-multipart==0.0.9

#Potom
transliterate==1.10.2
alembic==1.14.1
MarkupSafe==3.0.2
Mako==1.3.9
httptools==0.6.4
pyOpenSSL==25.0.0
pyyaml==6.0.2 
watchfiles==1.0.4 
websockets==15.0
deprecated==1.2.18 
limits==4.0.1 
packaging==24.2 
slowapi==0.1.9 
wrapt==1.17.2
#eshyo
# Jinja2==3.1.5 
# aiosmtplib==3.0.2 
# blinker==1.9.0 
# fastapi-mail==1.4.2 
# pydantic==2.10.6 
# pydantic-core==2.27.2 
# pydantic-settings==2.8.1
# === AI Integration ===
ollama==0.6.1                    # Клиент для Ollama (чат-бот)
httpx==0.28.1                    # HTTP-клиент (ollama использует его внутри)
email-validator==2.3.0           # Валидация email в Pydantic
aiohttp>=3.8.0